+++ scriptUtils.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const fs = require('node:fs');
+const { globSync }= require('glob');
 
 function getRedirectionsFilePath() {
     return path.resolve(__dirname + '/src/pages/redirects.json');
@@ -24,8 +25,30 @@ function writeRedirectionsFile(data) {
     fs.writeFileSync(redirectionsFilePath, JSON.stringify(redirectionsData));
 }
 
+function getMarkdownFiles() {
+    return globSync(__dirname + '/src/pages/**/*.md')
+        .map(f => path.resolve(f));
+}
+
+const getFindPatternForMarkdownFiles = (from) => `(\\[[^\\]]*]\\()(/|./)?(${from})(#[^\\()]*)?(\\))`;
+const getReplacePatternForMarkdownFiles = (to) => `$1$2${to}$4$5`;
+
+function replaceLinksInFile({ file, linkMap, getFindPattern, getReplacePattern }) {
+    let data = fs.readFileSync(file, 'utf8');
+    linkMap.forEach((to, from) => {
+        const find = getFindPattern(from);
+        const replace = getReplacePattern(to);
+        data = data.replaceAll(new RegExp(find, "gm"), replace);
+    });
+    fs.writeFileSync(file, data, 'utf-8');
+}
+
 module.exports = {
     getRedirectionsFilePath,
     readRedirectionsFile,
-    writeRedirectionsFile
+    writeRedirectionsFile,
+    getMarkdownFiles,
+    getFindPatternForMarkdownFiles,
+    getReplacePatternForMarkdownFiles,
+    replaceLinksInFile
 };
